Resolve "72-user_todo_owner"

See merge request KLINTez/fastapiLR!89

+++ application/todo/routes.py
@@ -405,7 +405,7 @@ async def todo_add(request: Request,
                    type: Annotated[str, Form()] = TodoTags.education.value,
                    source: Annotated[str, Form()] = Source.source_created.value,
                    details: Annotated[str, Form(max_length=500)] = None,
-                   fullname: Annotated[str, Form()] = Users.user1.value,
+                   fullname: Annotated[str, Form()] = "user",
                    date_creation: Annotated[date, Form()] = date.today(),
                    completed: Annotated[bool, Form()] = False,
                    date_completion: Annotated[date, Form()] = None,
@@ -419,7 +419,7 @@ async def todo_add(request: Request,
                            details=details,
                            type=type,
                            source=source,
-                           fullname=fullname,
+                           fullname=fullname if current_user.name == "admin" else current_user.name,
                            completed=completed,
                            date_creation=date_creation,
                            date_completion=date_completion)
@@ -471,6 +471,8 @@ async def todo_edit(
     """
     todo = database.query(models.Todo).filter(models.Todo.id == todo_id).first()
     if todo is not None and title is not None and title.replace(" ", "") != "":
+        if current_user.name != todo.fullname:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
         todo = database.query(models.Todo).filter(models.Todo.id == todo_id).first()
         logger.info(f"Editting todo: {todo}")
         todo.title = title
@@ -499,6 +501,8 @@ async def todo_delete(request: Request,
     todo = database.query(models.Todo).filter(models.Todo.id == todo_id).first()
     if todo is None:
         raise HTTPException(status_code=301)
+    if current_user.name != todo.fullname:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
     logger.info(f"Deleting todo: {todo}")
     database.delete(todo)
     database.commit()
@@ -513,10 +517,14 @@ async def todo_delete_all(request: Request,
                           ):
     """Delete all todos"""
     all_todos = database.query(models.Todo).all()
+        
     for todo in all_todos:
+        if current_user.name != todo.fullname:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
         await todo_delete(request=request,
                           todo_id=todo.id,
-                          database=database)
+                          database=database,
+                          current_user=current_user)
     return {"answer": "ok"}
 
 
@@ -544,7 +552,9 @@ async def todo_delete_range(request: Request,
 
     # todos = database.query(models.Todo).order_by(models.Todo.id.desc()).filter(models.Todo.type == type).all()
     todos = query.order_by(models.Todo.id.desc()).offset(start - 1).limit(end - start + 1)
-
+    for todo in todos:
+        if current_user.name != todo.fullname:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
     for todo in todos:
         database.delete(todo)
         deleting_todo(todo.id)
